refactor(use-overlay): `useOverlay`에서 `OverlayController`를 분리한다
